Separate interfaces to keep the public api clean
